Внес изменения в работу с БД
