+++ DataBase/Procedures.php
@@ -1,4 +1,5 @@
 <?php
+require_once('DataBase\SqlServerHelper.php');
 
 /**Список хранимых процедур из БД MySQL*/
 class Procedures
