+++ post.php
@@ -1,78 +1,25 @@
 <?php
 require_once('helpers.php');
 require_once('connection.php');
+require_once('DataBase\Procedures.php');
+require_once('DataBase\SqlFunctions.php');
 
 $is_auth = rand(0, 1);
 $user_name = 'Егор Толбаев'; // укажите здесь ваше имя
 $page_title = 'Readme: Публикация';
 
-/**
- * Обработка запросов
- *
- * @param mysqli string Строка соединения
- * @param $stringSQL string Запрос
- *
- * @return  array $result
- */
-function requestHandler(mysqli $connect, string $stringSQL): array
-{
-    $result = mysqli_query($connect, $stringSQL);
-    if (!$result) {
-        exit("Ошибка MySQL: " . mysqli_error($connect));
-    }
-    return mysqli_fetch_all($result, MYSQLI_ASSOC);
-}
+$procedures = new Procedures();
+$sqlFunctions = new SqlFunctions();
 
 $post_id = isset($_GET['post_id']) ? $_GET['post_id'] : null;
 
-//region Функции для получения информации из БД
-/**
- * Полная информация по посту  по его id
- * @param mysqli $connect
- * @param string $post_id
- * @return array|null
- */
-function GetPostById(mysqli $connect, string $post_id)
-{
-    $sql = "SELECT p.*, ct.icon_type, u.avatar, u.login,
-    IFNULL((SELECT COUNT(*) FROM likes l WHERE l.post_id = p.id), 0) AS likes_count,
-    IFNULL((SELECT COUNT(*) FROM comments com WHERE com.post_id = p.id), 0) AS comments_count,
-    IFNULL((SELECT COUNT(*) FROM subscriptions sub WHERE sub.user_id = p.user_id), 0) AS subscribers_count
-    FROM posts p
-    JOIN users u ON p.user_id = u.id
-    JOIN content_type ct ON p.type_id = ct.id
-    LEFT JOIN likes l ON l.post_id = p.id
-    LEFT JOIN comments com ON com.post_id = p.id
-    LEFT JOIN subscriptions sub ON sub.user_id = p.user_id
-    WHERE p.id = $post_id
-    GROUP BY l.id";
-    $result = requestHandler($connect, $sql);
-    return empty($result) ? NULL : $result;
-}
-
-/**
- * Получаем количество постов пользователя по id
- * @param mysqli $connect
- * @param string $user_id идентификатор пользователя
- * @return array
- */
-function GetUserPostsCount(mysqli $connect, string $user_id)
-{
-    $sql = "SELECT COUNT(*) AS posts_count FROM posts p
-    WHERE p.user_id = $user_id";
-    $result = requestHandler($connect, $sql);
-    return empty($result) ? NULL : $result;
-}
-
-//endregion
-
-$post = GetPostById($mainConnection, $post_id)[0];
+$post = $sqlFunctions->GetPostById($mainConnection, $post_id)[0];
 
 if (!$post) {
     $page_content = include_template('Error-404.php', []);
 } else {
     $post_content = include_template("post-types/post-{$post['icon_type']}.php", ['post' => $post]);
-    $posts_count = GetUserPostsCount($mainConnection, $post['user_id']);
+    $posts_count = $sqlFunctions->GetUserPostsCount($mainConnection, $post['user_id']);
     $page_content = include_template('post-show.php', [
         'post_content' => $post_content,
         'post' => $post,
